Create file for put method
